perf: fuse DeepSeek-V3.2 DSA indexer
